refactor(mocks): extract schedules module from mockApi (#991 M3)
